third_party/blink/web_tests/wpt_internal/document-transition/mix-blend-mode-only-on-transition.html

<!DOCTYPE html>
<html class=reftest-wait>
<title>SET: Blend modes are set up only in paired transitions</title>
<link rel="help" href="https://github.com/WICG/shared-element-transitions">
<link rel="author" href="mailto:vmpstr@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
div {
  width: 100px;
  height: 100px;
  background: blue;
  contain: paint;
}
.tagone { page-transition-tag: one }
.tagtwo { page-transition-tag: two }
.tagthree { page-transition-tag: three }
</style>

<div id=first class=tagone></div>
<div id=second></div>
<div id=third class=tagthree></div>

<script>
async_test(t => {
  let transition = document.createDocumentTransition();
  transition.start(() => {
    first.classList.toggle("tagone");
    second.classList.toggle("tagtwo");
    requestAnimationFrame(() => {
      requestAnimationFrame(() => {
        t.step(() => assert_equals(getComputedStyle(document.documentElement, "::page-transition-image-wrapper(one)").isolation, "auto"));
        t.step(() => assert_equals(getComputedStyle(document.documentElement, "::page-transition-outgoing-image(one)").mixBlendMode, "normal"));
        t.step(() => assert_equals(getComputedStyle(document.documentElement, "::page-transition-image-wrapper(two)").isolation, "auto"));
        t.step(() => assert_equals(getComputedStyle(document.documentElement, "::page-transition-incoming-image(two)").mixBlendMode, "normal"));
        t.step(() => assert_equals(getComputedStyle(document.documentElement, "::page-transition-image-wrapper(three)").isolation, "isolate"));
        t.step(() => assert_equals(getComputedStyle(document.documentElement, "::page-transition-outgoing-image(three)").mixBlendMode, "plus-lighter"));
        t.step(() => assert_equals(getComputedStyle(document.documentElement, "::page-transition-incoming-image(three)").mixBlendMode, "plus-lighter"));
        t.done();
      });
    });
  });
}, "Blend modes are set up on paired transitions");
</script>
